Added Tutoral Base and Event Page Priority Tutorial Map
Added a basic RPGMaker game the anyone following my tutorials can
download and run my tutoiral maps on.
To save on constantly re-uploading the same files a single core game was
generated. The user needs to swap the map001 file to get the code used
in the tutorial.
The map for a tutorial on Event page priority is complete and after this
commit it will be uploaded with a link to this repository.